+++ cpskin/core/upgradehandlers.py
@@ -11,6 +11,7 @@ from cpskin.core.setuphandlers import addAutoPlaySliderToRegistry
 from cpskin.core.setuphandlers import addCityNameToRegistry
 from cpskin.core.setuphandlers import addCollapseMinisiteMenuToRegistry
 from cpskin.core.setuphandlers import addFooterSitemapToRegistry
+from cpskin.core.setuphandlers import addContentClassesToRegistry
 from cpskin.core.setuphandlers import addDescriptionOnThemesOptionToRegistry
 from cpskin.core.setuphandlers import addLoadPageMenuToRegistry
 from cpskin.core.setuphandlers import addMediaViewletOptionsToRegistry
@@ -191,6 +192,9 @@ def upgrade_registry_for_minisite_menu(context):
 
 def upgrade_registry_for_show_footer_sitemap(context):
     addFooterSitemapToRegistry()
+
+def upgrade_registry_for_content_classes(context):
+    addContentClassesToRegistry()
 
 def update_types(context):
     context.runImportStepFromProfile('profile-cpskin.core:default', 'typeinfo')