+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 0.13.12 (unreleased)
 --------------------
 
+- Add body classes for header and content blocs : WEB-3027
+  [laulaz]
+
 - Add body class for minisite (independant of mode)
   [laulaz]